Change gas from u128 to u64 (#890)
* Revert "Revert builtin cost rework (#874)"

This reverts commit 355c250f37cf0977ef2776b1aae2cb2e87c9da3d.

* x

* fix gas

* add paralel test

* Add entry point info to AotContractExecutor (#889)

* store entry point info, optimize abi storage

* use felt

* update hash

* upd replay

* try with thread local

* try to fix with thread local in runtime2

* fix

* upd ref ci

* upd replay

* fix macos

* try to fix macos

* try to fix bench

* fix

* try to fix bench ci

* save space

* upd rev

* try to fix broken pipe

* try to fix bench on ci

* remove all features

* add cache key

* update replay rev

* fix cargo.lock

* Revert "update replay rev"

This reverts commit 40f1acc5c10687ed6c43a0a7dbd09c581ca10ec6.

* fix cargo.lock

* works with leaks

* fix

* upd replay

* fix bench

* Reorder deps

* start porting gas to u64

* u64 on syscalls

* fix

* Fix benches.

* clippy

* fix gas refund

* fmt

* fix

* fix get gas

* update resource bounds

* upd ci

* upd replay

* fix ci

* try

* fix align

* fix bench pr

* better fix for u256

* upd replay

* upd replay

* fix

* remove dbg!

* upd replay

* fix

* Revert conversion of `ResourceBounds` field type.

* Update sequencer revision (for CI runs).

* Update replay dependency.

* read invoke data end pad

* upd ci

---------

+++ tests/tests/felt252.rs
@@ -71,7 +71,7 @@ proptest! {
             program,
             "run_test",
             &[Value::Felt252(a), Value::Felt252(b)],
-            Some(DEFAULT_GAS as u128),
+            Some(DEFAULT_GAS),
             Option::<DummySyscallHandler>::None,
         );
 
@@ -97,7 +97,7 @@ proptest! {
             program,
             "run_test",
             &[Value::Felt252(a), Value::Felt252(b)],
-            Some(DEFAULT_GAS as u128),
+            Some(DEFAULT_GAS),
             Option::<DummySyscallHandler>::None,
         );
 
@@ -123,7 +123,7 @@ proptest! {
             program,
             "run_test",
             &[Value::Felt252(a), Value::Felt252(b)],
-            Some(DEFAULT_GAS as u128),
+            Some(DEFAULT_GAS),
             Option::<DummySyscallHandler>::None,
         );
 
@@ -149,7 +149,7 @@ proptest! {
             program,
             "run_test",
             &[Value::Felt252(a), Value::Felt252(b)],
-            Some(DEFAULT_GAS as u128),
+            Some(DEFAULT_GAS),
             Option::<DummySyscallHandler>::None,
         );
